+++ common/placeholders/src/main/java/io/knotx/server/common/placeholders/JsonPlaceholderSubstitutor.java
@@ -21,7 +21,7 @@ import org.apache.commons.lang3.StringUtils;
 
 import io.vertx.core.json.JsonObject;
 
-public class JsonPlaceholderSubstitutor implements PlaceholderSubstitutor<JsonObject> {
+class JsonPlaceholderSubstitutor implements PlaceholderSubstitutor<JsonObject> {
   @Override
   public String getValue(JsonObject source, String placeholder) {
     return getString(StringUtils.substringAfter(placeholder, "."), source);
