+++ common/placeholders/src/main/java/io/knotx/server/common/placeholders/SlingUriInfo.java
@@ -15,7 +15,7 @@
  */
 package io.knotx.server.common.placeholders;
 
-public class SlingUriInfo {
+class SlingUriInfo {
 
   private String path;
