--- common/placeholders/src/main/java/io/knotx/server/common/placeholders/configuration/SourceDefinition.java
@@ -1,43 +1,0 @@
-/*
- * Copyright (C) 2019 Knot.x Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.knotx.server.common.placeholders.configuration;
-
-import java.util.Set;
-
-public class SourceDefinition<T> {
-
-  private final Class<T> clazz;
-  private final T source;
-  private final Set<String> prefixes;
-
-  public SourceDefinition(Class<T> clazz, T source, Set<String> prefixes) {
-    this.clazz = clazz;
-    this.source = source;
-    this.prefixes = prefixes;
-  }
-
-  public T getSource() {
-    return source;
-  }
-
-  public Set<String> getPrefixes() {
-    return prefixes;
-  }
-
-  public Class<T> getSourceClass() {
-    return clazz;
-  }
-}
